refactoring: model, enum들 domain 모듈로 이전
