+++ domain/src/main/java/com/fintory/domain/alarm/model/Alarm.java
@@ -1,0 +1,28 @@
+package com.fintory.domain.alarm.model;
+
+
+
+import com.fintory.domain.child.model.Child;
+import com.fintory.domain.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name="alarm")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Alarm extends BaseEntity {
+
+    private String type;
+
+    private String content;
+
+    @Column(name="is_read")
+    private boolean isRead;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="child_id")
+    private Child child;
+}
