+++ domain/src/main/java/com/fintory/domain/mapping/MappingStatus.java
@@ -1,0 +1,5 @@
+package com.fintory.domain.mapping;
+
+public enum MappingStatus {
+    INVITED, REJECTED,DISCONNECTED
+}
