+++ domain/src/main/java/com/fintory/domain/mapping/UserBadge.java
@@ -1,0 +1,25 @@
+package com.fintory.domain.mapping;
+
+
+import com.fintory.domain.badge.model.Badge;
+import com.fintory.domain.child.model.Child;
+import com.fintory.domain.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name="user_badges")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserBadge extends BaseEntity {
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="child_id")
+    private Child child;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="badge_id")
+    private Badge badge;
+}
