+++ domain/src/main/java/com/fintory/domain/parent/model/Role.java
@@ -1,0 +1,6 @@
+package com.fintory.domain.parent.model;
+
+public enum Role {
+    mother,
+    father
+}
