+++ domain/src/main/java/com/fintory/domain/portfolio/model/MarketType.java
@@ -1,0 +1,5 @@
+package com.fintory.domain.portfolio.model;
+
+public enum MarketType {
+    DOMESTIC, OVERSEAS
+}
