+++ domain/src/main/java/com/fintory/domain/portfolio/model/Status.java
@@ -1,0 +1,5 @@
+package com.fintory.domain.portfolio.model;
+
+public enum Status {
+    PENDING, COMPLETED, CANCELED, FAILED
+}
