+++ domain/src/main/java/com/fintory/domain/stock/model/IntervalType.java
@@ -1,0 +1,5 @@
+package com.fintory.domain.stock.model;
+
+public enum IntervalType {
+    daily, monthly, yearly
+}
