--- src/main/java/com/campuspick/fintory/domain/child/domain/entity/MappingStatus.java
@@ -1,5 +1,0 @@
-package com.campuspick.fintory.domain.child.domain.entity;
-
-public enum MappingStatus {
-    INVITED, REJECTED,DISCONNECTED
-}
