--- src/main/java/com/campuspick/fintory/domain/child/domain/entity/ParentChildMapping.java
@@ -1,33 +1,0 @@
-package com.campuspick.fintory.domain.child.domain.entity;
-
-
-import com.campuspick.fintory.domain.parent.domain.entity.Parent;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@NoArgsConstructor
-@Getter
-@Table(name="parent_child_mappings")
-public class ParentChildMapping {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private MappingStatus mappingStatus;
-
-    @Column(name="invited_at")
-    private LocalDateTime invitedAt;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="parent_id")
-    private Parent parent;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="child_id")
-    private Child child;
-}
