--- src/main/java/com/campuspick/fintory/domain/portfolio/domain/entity/Status.java
@@ -1,5 +1,0 @@
-package com.campuspick.fintory.domain.portfolio.domain.entity;
-
-public enum Status {
-    PENDING, COMPLETED, CANCELED, FAILED
-}
