--- src/main/java/com/campuspick/fintory/global/exception/ErrorCode.java
@@ -1,22 +1,0 @@
-package com.campuspick.fintory.global.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public enum ErrorCode {
-
-    VALIDATION_FAIL(HttpStatus.BAD_REQUEST, "VALIDATION_FAIL", "입력값이 유효하지 않습니다."),
-    USER_NOT_FOUND(HttpStatus.BAD_REQUEST, "USER_NOT_FOUND", "유저를 찾을 수 없습니다."),
-    NEWS_NOT_FOUND(HttpStatus.BAD_REQUEST, "NEWS_NOT_FOUND", "뉴스를 찾을 수 없습니다");
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-
-    ErrorCode(HttpStatus httpStatus, String code, String message) {
-        this.httpStatus = httpStatus;
-        this.code = code;
-        this.message = message;
-    }
-}
